+++ Source/Nett.AspNet/ProviderDictionaryConverter.cs
@@ -19,6 +19,7 @@ namespace Nett.AspNet
             {
                 switch (r.Value)
                 {
+                    case TomlTableArray ta: ProcessTableArray(dict, ta, FullKey(r)); break;
                     case TomlTable t: ProcessTable(dict, t, keyPrefix + r.Key + ":"); break;
                     case TomlArray a: ProcessArray(dict, a, FullKey(r)); break;
                     case TomlBool b: AddEntry(r, b.Value.ToString(CultureInfo.InvariantCulture)); break;
@@ -43,18 +44,27 @@ namespace Nett.AspNet
                 => keyPrefix + row.Key;
         }
 
+        private static void ProcessTableArray(Dictionary<string, string> dict, TomlTableArray tableArray, string keyPrefix)
+        {
+            for (int i = 0; i < tableArray.Items.Count; i++)
+            {
+                TomlTable table = tableArray.Items[i];
+                ProcessTable(dict, table, $"{keyPrefix}:{i}:");
+            }
+        }
+
         private static void ProcessArray(Dictionary<string, string> dict, TomlArray array, string fullKey)
         {
             for (int i = 0; i < array.Items.Length; i++)
             {
                 if (array.Items[i] is TomlArray a)
                 {
-                    throw new InvalidOperationException(
-                        $"AspNet provider cannot handle jagged arrays, only simple arrays are supported." +
-                        $"The arrays key is '{fullKey}'.");
+                    ProcessArray(dict, a, $"{fullKey}:{i}");
                 }
-
-                dict[$"{fullKey}:{i}"] = array.Items[i].ToString();
+                else
+                {
+                    dict[$"{fullKey}:{i}"] = array.Items[i].ToString();
+                }
             }
         }
     }